// src/app/generated/cite.api/model/evaluationPermission.ts
/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type EvaluationPermission = 'ViewEvaluation' | 'EditEvaluation' | 'ManageEvaluation' | 'ExecuteEvaluation' | 'ObserveEvaluation' | 'ParticipateInEvaluation';

export const EvaluationPermission = {
    ViewEvaluation: 'ViewEvaluation' as EvaluationPermission,
    EditEvaluation: 'EditEvaluation' as EvaluationPermission,
    ManageEvaluation: 'ManageEvaluation' as EvaluationPermission,
    ExecuteEvaluation: 'ExecuteEvaluation' as EvaluationPermission,
    ObserveEvaluation: 'ObserveEvaluation' as EvaluationPermission,
    ParticipateInEvaluation: 'ParticipateInEvaluation' as EvaluationPermission
};