// src/app/generated/cite.api/model/evaluationRole.ts
/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EvaluationPermission } from './evaluationPermission';


export interface EvaluationRole { 
    id?: string;
    name?: string | null;
    allPermissions?: boolean;
    permissions?: Array<EvaluationPermission> | null;
}